refactor(fe): co-locate each MCP trust rule with the route that uses it
`mcpConfig.ts` was carrying two near-identical rules that differ by one
branch, next to the canister read/write helpers everyone imports. Nothing
stopped a call site reaching for the wrong one.

Each now lives beside its only caller: `trustedUrl` — what the identity
trusts right now — with the Settings route, and `isOriginTrusted` — whether
a new connect may proceed — with the `/mcp` consent route. Their tests moved
with them. `mcpConfig.ts` is left with the shared config type, `originOf`,
and the canister accessors.
